@@ -4,15 +4,13 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "react": "^17.0.2",
-    "react-animate-on-scroll": "^2.1.5",
-    "react-dom": "^17.0.2",
-    "react-ga": "^3.3.0",
+    "react": "^18.0.0",
+    "react-animate-on-scroll": "^2.1.9",
+    "react-dom": "^18.0.0",
+    "react-ga": "^3.3.1",
     "react-scripts": "^5.0.1",
-    "web-vitals": "^1.1.2"
+    "web-vitals": "^5.0.0",
+    "react-router": "^7.6.0"
   },
   "scripts": {
     "predeploy": "npm run build",
@@ -41,7 +39,6 @@
     ]
   },
   "devDependencies": {
-    "gh-pages": "^6.3.0",
-    "react-router-dom": "^5.3.0"
+    "gh-pages": "^6.3.0"
   }
 }
